+++ tests/unit/V1/async_db/test_cursor.py
@@ -608,13 +608,13 @@ async def test_cursor_skip_parse(
     httpx_mock.add_callback(query_callback, url=query_url)
 
     with patch(
-        "firebolt.common._types.StatementFormatter.split_format_sql"
+        "firebolt.common.statement_formatter.StatementFormatter.split_format_sql"
     ) as split_format_sql_mock:
         await cursor.execute("non-an-actual-sql")
         split_format_sql_mock.assert_called_once()
 
     with patch(
-        "firebolt.common._types.StatementFormatter.split_format_sql"
+        "firebolt.common.statement_formatter.StatementFormatter.split_format_sql"
     ) as split_format_sql_mock:
         await cursor.execute("non-an-actual-sql", skip_parsing=True)
         split_format_sql_mock.assert_not_called()
