+++ tests/unit/async_db/test_cursor.py
@@ -539,13 +539,13 @@ async def test_cursor_skip_parse(
     mock_query()
 
     with patch(
-        "firebolt.common._types.StatementFormatter.split_format_sql"
+        "firebolt.common.statement_formatter.StatementFormatter.split_format_sql"
     ) as split_format_sql_mock:
         await cursor.execute("non-an-actual-sql")
         split_format_sql_mock.assert_called_once()
 
     with patch(
-        "firebolt.common._types.StatementFormatter.split_format_sql"
+        "firebolt.common.statement_formatter.StatementFormatter.split_format_sql"
     ) as split_format_sql_mock:
         await cursor.execute("non-an-actual-sql", skip_parsing=True)
         split_format_sql_mock.assert_not_called()
